Working on encorporating game window to qt interface
